+++ examples/20_embedded_basic.json
