+++ examples/32_array_embed.json
