+++ lib/generator.js
