+++ lib/inserter.js
