+++ test/configArray.test.js
@@ -1,4 +1,4 @@
-var helper = require('./helper');
+var helpers = require('./helpers');
 var assert = require('assert');
 
 describe('configurable arrays', function() {
@@ -20,7 +20,7 @@ describe('configurable arrays', function() {
         }
       }
     };
-    helper.getResults(opts, function (err, items) {
+    helpers.getResults(opts, function (err, items) {
       if (err) return done(err);
       res.items = items;
       done();
