+++ test/types.test.js
@@ -1,4 +1,4 @@
-var helper = require('./helper');
+var helpers = require('./helpers');
 var assert = require('assert');
 
 describe('generate different types of data', function() {
@@ -26,7 +26,7 @@ describe('generate different types of data', function() {
         },
       }
     };
-    helper.getResults(opts, function (err, items) {
+    helpers.getResults(opts, function (err, items) {
       if (err) return done(err);
       res.item = items[0];
       done();
